Refactor library organisation
