+++ src/LibUHD/functions.jl
@@ -1,229 +1,3 @@
-module LibUHD
-
-# ----------------------------------------------------
-# --- Dependencies 
-# ---------------------------------------------------- 
-using USRPHardwareDriver_jll
-# export USRPHardwareDriver_jll
-using CEnum
-
-# ----------------------------------------------------
-# --- Constant
-# ---------------------------------------------------- 
-const UHD_VERSION_ABI_STRING = "4.1.0"
-const UHD_VERSION = 4010099
-
-
-# ----------------------------------------------------
-# --- Enumeration
-# ---------------------------------------------------- 
-@cenum uhd_error::UInt32 begin
-    UHD_ERROR_NONE = 0
-    UHD_ERROR_INVALID_DEVICE = 1
-    UHD_ERROR_INDEX = 10
-    UHD_ERROR_KEY = 11
-    UHD_ERROR_NOT_IMPLEMENTED = 20
-    UHD_ERROR_USB = 21
-    UHD_ERROR_IO = 30
-    UHD_ERROR_OS = 31
-    UHD_ERROR_ASSERTION = 40
-    UHD_ERROR_LOOKUP = 41
-    UHD_ERROR_TYPE = 42
-    UHD_ERROR_VALUE = 43
-    UHD_ERROR_RUNTIME = 44
-    UHD_ERROR_ENVIRONMENT = 45
-    UHD_ERROR_SYSTEM = 46
-    UHD_ERROR_EXCEPT = 47
-    UHD_ERROR_BOOSTEXCEPT = 60
-    UHD_ERROR_STDEXCEPT = 70
-    UHD_ERROR_UNKNOWN = 100
-end
-
-@cenum uhd_rx_metadata_error_code_t::UInt32 begin
-    UHD_RX_METADATA_ERROR_CODE_NONE = 0
-    UHD_RX_METADATA_ERROR_CODE_TIMEOUT = 1
-    UHD_RX_METADATA_ERROR_CODE_LATE_COMMAND = 2
-    UHD_RX_METADATA_ERROR_CODE_BROKEN_CHAIN = 4
-    UHD_RX_METADATA_ERROR_CODE_OVERFLOW = 8
-    UHD_RX_METADATA_ERROR_CODE_ALIGNMENT = 12
-    UHD_RX_METADATA_ERROR_CODE_BAD_PACKET = 15
-end
-
-@cenum uhd_async_metadata_event_code_t::UInt32 begin
-    UHD_ASYNC_METADATA_EVENT_CODE_BURST_ACK = 1
-    UHD_ASYNC_METADATA_EVENT_CODE_UNDERFLOW = 2
-    UHD_ASYNC_METADATA_EVENT_CODE_SEQ_ERROR = 4
-    UHD_ASYNC_METADATA_EVENT_CODE_TIME_ERROR = 8
-    UHD_ASYNC_METADATA_EVENT_CODE_UNDERFLOW_IN_PACKET = 16
-    UHD_ASYNC_METADATA_EVENT_CODE_SEQ_ERROR_IN_BURST = 32
-    UHD_ASYNC_METADATA_EVENT_CODE_USER_PAYLOAD = 64
-end
-
-@cenum uhd_sensor_value_data_type_t::UInt32 begin
-    UHD_SENSOR_VALUE_BOOLEAN = 98
-    UHD_SENSOR_VALUE_INTEGER = 105
-    UHD_SENSOR_VALUE_REALNUM = 114
-    UHD_SENSOR_VALUE_STRING = 115
-end
-
-@cenum uhd_tune_request_policy_t::UInt32 begin
-    UHD_TUNE_REQUEST_POLICY_NONE = 78
-    UHD_TUNE_REQUEST_POLICY_AUTO = 65
-    UHD_TUNE_REQUEST_POLICY_MANUAL = 77
-end
-
-@cenum uhd_stream_mode_t::UInt32 begin
-    UHD_STREAM_MODE_START_CONTINUOUS = 97
-    UHD_STREAM_MODE_STOP_CONTINUOUS = 111
-    UHD_STREAM_MODE_NUM_SAMPS_AND_DONE = 100
-    UHD_STREAM_MODE_NUM_SAMPS_AND_MORE = 109
-end
-
-@cenum uhd_log_severity_level_t::UInt32 begin
-    UHD_LOG_LEVEL_TRACE = 0
-    UHD_LOG_LEVEL_DEBUG = 1
-    UHD_LOG_LEVEL_INFO = 2
-    UHD_LOG_LEVEL_WARNING = 3
-    UHD_LOG_LEVEL_ERROR = 4
-    UHD_LOG_LEVEL_FATAL = 5
-end
-# ----------------------------------------------------
-# --- Structures
-# ---------------------------------------------------- 
-mutable struct uhd_rx_metadata_t end
-
-mutable struct uhd_tx_metadata_t end
-
-mutable struct uhd_async_metadata_t end
-
-const uhd_rx_metadata_handle = Ptr{uhd_rx_metadata_t}
-
-const uhd_tx_metadata_handle = Ptr{uhd_tx_metadata_t}
-
-const uhd_async_metadata_handle = Ptr{uhd_async_metadata_t}
-
-struct uhd_range_t
-    start::Cdouble
-    stop::Cdouble
-    step::Cdouble
-end
-
-mutable struct uhd_meta_range_t end
-
-const uhd_meta_range_handle = Ptr{uhd_meta_range_t}
-
-
-mutable struct uhd_sensor_value_t end
-
-const uhd_sensor_value_handle = Ptr{uhd_sensor_value_t}
-
-
-
-mutable struct uhd_string_vector_t end
-
-const uhd_string_vector_handle = Ptr{uhd_string_vector_t}
-
-
-struct uhd_tune_request_t
-    target_freq::Cdouble
-    rf_freq_policy::uhd_tune_request_policy_t
-    rf_freq::Cdouble
-    dsp_freq_policy::uhd_tune_request_policy_t
-    dsp_freq::Cdouble
-    args::Ptr{Cchar}
-end
-
-struct uhd_tune_result_t
-    clipped_rf_freq::Cdouble
-    target_rf_freq::Cdouble
-    actual_rf_freq::Cdouble
-    target_dsp_freq::Cdouble
-    actual_dsp_freq::Cdouble
-end
-
-struct uhd_usrp_rx_info_t
-    mboard_id::Ptr{Cchar}
-    mboard_name::Ptr{Cchar}
-    mboard_serial::Ptr{Cchar}
-    rx_id::Ptr{Cchar}
-    rx_subdev_name::Ptr{Cchar}
-    rx_subdev_spec::Ptr{Cchar}
-    rx_serial::Ptr{Cchar}
-    rx_antenna::Ptr{Cchar}
-end
-
-struct uhd_usrp_tx_info_t
-    mboard_id::Ptr{Cchar}
-    mboard_name::Ptr{Cchar}
-    mboard_serial::Ptr{Cchar}
-    tx_id::Ptr{Cchar}
-    tx_subdev_name::Ptr{Cchar}
-    tx_subdev_spec::Ptr{Cchar}
-    tx_serial::Ptr{Cchar}
-    tx_antenna::Ptr{Cchar}
-end
-
-
-mutable struct uhd_dboard_eeprom_t end
-
-const uhd_dboard_eeprom_handle = Ptr{uhd_dboard_eeprom_t}
-
-mutable struct uhd_mboard_eeprom_t end
-
-const uhd_mboard_eeprom_handle = Ptr{uhd_mboard_eeprom_t}
-
-struct uhd_subdev_spec_pair_t
-    db_name::Ptr{Cchar}
-    sd_name::Ptr{Cchar}
-end
-
-mutable struct uhd_subdev_spec_t end
-
-const uhd_subdev_spec_handle = Ptr{uhd_subdev_spec_t}
-
-struct uhd_usrp_register_info_t
-    bitwidth::Csize_t
-    readable::Bool
-    writable::Bool
-end
-
-struct uhd_stream_args_t
-    cpu_format::Ptr{Cchar}
-    otw_format::Ptr{Cchar}
-    args::Ptr{Cchar}
-    channel_list::Ptr{Csize_t}
-    n_channels::Cint
-end
-
-
-
-struct uhd_stream_cmd_t
-    stream_mode::uhd_stream_mode_t
-    num_samps::Csize_t
-    stream_now::Bool
-    time_spec_full_secs::Int64
-    time_spec_frac_secs::Cdouble
-end
-
-mutable struct uhd_rx_streamer end
-
-mutable struct uhd_tx_streamer end
-
-const uhd_rx_streamer_handle = Ptr{uhd_rx_streamer}
-
-const uhd_tx_streamer_handle = Ptr{uhd_tx_streamer}
-
-mutable struct uhd_usrp end
-
-const uhd_usrp_handle = Ptr{uhd_usrp}
-
-
-mutable struct uhd_usrp_clock end
-
-const uhd_usrp_clock_handle = Ptr{uhd_usrp_clock}
-# ----------------------------------------------------
-# --- Function calls
-# ---------------------------------------------------- 
 function uhd_get_last_error(error_out, strbuffer_len)
     ccall((:uhd_get_last_error, libuhd), uhd_error, (Ptr{Cchar}, Csize_t), error_out, strbuffer_len)
 end
@@ -1120,21 +894,4 @@ function uhd_set_thread_priority(priority, realtime)
     ccall((:uhd_set_thread_priority, libuhd), uhd_error, (Cfloat, Bool), priority, realtime)
 end
 
-# Skipping MacroDefinition: UHD_EXPORT __attribute__ ( ( visibility ( "default" ) ) )
 
-# Skipping MacroDefinition: UHD_IMPORT __attribute__ ( ( visibility ( "default" ) ) )
-
-# Skipping MacroDefinition: UHD_INLINE inline __attribute__ ( ( always_inline ) )
-
-# Skipping MacroDefinition: UHD_DEPRECATED __attribute__ ( ( deprecated ) )
-
-# const UHD_API = UHD_IMPORT
-# exports
-const PREFIXES = ["UHD_","uhd_"]
-for name in names(@__MODULE__; all=true), prefix in PREFIXES
-    if startswith(string(name), prefix)
-        @eval export $name
-    end
-end
-
-end # module
